Section 95 - Validate JwtToken (Custom Validate)

+++ Common/Common.AspNetCore/ClaimUtils.cs
@@ -4,11 +4,11 @@ namespace Common.AspNetCore;
 
 public static class ClaimUtils
 {
-    public static int GetUserId(this ClaimsPrincipal principal)
+    public static long GetUserId(this ClaimsPrincipal principal)
     {
         if (principal == null)
             throw new ArgumentNullException(nameof(principal));
 
-        return Convert.ToInt32(principal.FindFirst(ClaimTypes.NameIdentifier)?.Value);
+        return Convert.ToInt64(principal.FindFirst(ClaimTypes.NameIdentifier)?.Value);
     }
 }
